edits

+++ src/BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
-    public int calculate (int kg, double m){
+    public  double calculate (int kg, double m){
         double result=m*m;
         double result2=kg/result;
-        return (int) result2;
+        return result2;
     }
 }
